Raise errors on undefined variables in templates
 ## Summary
Jinja2 is a very forgiving templating language, by default (in Flask).
If a variable is undefined and a template tries to access that variable,
it is resolved as `none` by default - and no warnings/errors are
generated. This makes it very easy to accidentally typo variable names,
or reference variables that haven't been passed into the template at
all, and the only way to notice this is through manual QA.

By enabling `StrictUndefined` mode, we enforce handling templates much
more explicitly. Any attempt to access a variable that is not defined
for the template will raise an `UndefinedError`. This will cause
templates to 'fail fast', which is a much easier way of detecting
configuration/implementation errors in templates. It also makes our unit
tests around rendering pages more likely to warn us of errors.

This change will result in a number of template rendering failures by
default, so will follow with a set of separate commits which address the
issues which arise.