`static_mode` declared for all templates
 ## Summary
Add `static_mode` as a config option for an app (False by default). Also
adds `static_mode` as a default variable exposed to all templates,
derived from the app's config or the request's query params. This allows
all templates to safely check whether or not they should be rendered in
static mode.