Refactor measure page nav links
 ## Summary
On measure pages (CMS-only), we include extra navigation links to edit
or export the measure. These are currently baked into the header,
meaning that every time the template is rendered we are checking for an
arbitrary `measure_page` variable - if present, we try to render extra
links.

In reality, we only want to display the extra links when _actually_
showing a measure page, so these additional nav links feel better
included in the `measure.html` template. This patch removes the nav
links from the shared `_header.html` partial template and replaces them
with a new `additionalNavLinks` block, which `measure.html` can then
hook into.