Always pass `new` parameter to edit measure page
 ## Summary
This template needs to know whether or not the user is editing an
existing measure or creating a new measure; it does this through the
`new` variable. Given this is a tightly-scoped variable (i.e. only
checked for in the `edit_measure_page.html` template), we should always
pass it in - whether it's true or false.