Custom `get_csrf_token` filter
 ## Summary
We render csrf fields in forms using the `csrf_token` attribute on WTF
forms. These are added dynamically if the config option
`WTF_CSRF_ENABLED` is on. If this is disabled (as during test runs), the
field is not present, which causes templates to fail to render. The
custom jinja2 filter will return either the CSRF token field or a blank
string depending on the value of the config option.

This feels slightly hacky, because it's technically code only added for
the sake of tests - but I think the overall benefit of StrictUndefined
is sufficient to offset this slight smell.