Create `AnonymousUser` class
 ## Summary
Some of our templates need to access the current user to check whether
said user is authorised to view certain pages/parts of pages. Anonymous
users do not have these attributes, so this patch creates an
`AnonymousUser` class that exposes the same interface as the main `User`
class, making the templates compatible with guests as well as logged-in
users.