Migration for new `DataSource` table
 ## Summary
Add a structural migration that implements the new `DataSource` table,
intended to deprecate and later remove a number of fields on the
existing `Page` model.

This retirement will happen across a number of stages:
1) Creating the new `DataSource` structural model.
2) Running a data migration to copy data from the `Page` table to the
`DataSource` model; at the same time, update views to start referencing
the `DataSource` table for future reads and writes.
3) Drop the existing data source columns from the `Page` table.

 ## Ticket
https://trello.com/c/yV12Lu7Z/949